lesson_10/requirements.txt
requests
pytest
sqlalchemy
psycopg2-binary
allure-pytest